Refactor moving and reodering
