+++ lib/Dialect/TTNN/Analysis/MatmulProgramConfig.cpp
@@ -9,6 +9,7 @@
 #include "ttmlir/Dialect/TTNN/Types/Types.h"
 #include "ttmlir/Dialect/TTNN/Utils/Utils.h"
 
+#include "mlir/IR/BuiltinTypes.h"
 #include "llvm/ADT/TypeSwitch.h"
 
 namespace mlir::tt::ttnn {
@@ -298,6 +299,161 @@ generateMatmulProgramConfig(Operation *op, TTNNLayoutAttr outputLayout) {
   return generateMatmul1DProgramConfig(ctx, Mt, Nt, Kt, outputLayout,
                                        outputMemLayout, fusedActivation,
                                        maxSubblockSize, fuseBatch);
+}
+
+// ============================================================================
+// DRAM-sharded matmul config generation
+// ============================================================================
+
+static constexpr int64_t kTileSize = 32;
+
+static int64_t padToDRAMBanks(int64_t n, int64_t numBanks) {
+  int64_t lcm = kTileSize * numBanks;
+  return ((n + lcm - 1) / lcm) * lcm;
+}
+
+std::optional<DRAMShardParams>
+computeShardParams(int64_t M, int64_t K, int64_t N, int64_t numBanks,
+                   int64_t numIn0Cores, int64_t numOutCores,
+                   ttcore::DataType weightDataType, int64_t l1Available) {
+  DRAMShardParams p;
+  p.K = K;
+  p.N = N;
+  p.M = M;
+  p.numBanks = numBanks;
+  p.numIn0Cores = numIn0Cores;
+  p.numOutCores = numOutCores;
+  p.nPadded = padToDRAMBanks(N, numBanks);
+  p.shardH = K;
+  p.shardW = p.nPadded / numBanks;
+  p.kTiles = K / kTileSize;
+  p.shardWTiles = p.shardW / kTileSize;
+  p.perCoreM = M / kTileSize;
+  p.perCoreN = (N / kTileSize + numOutCores - 1) / numOutCores; // div_up
+  p.in0ShardW = K / numIn0Cores;
+  p.weightDataType = weightDataType;
+
+  static constexpr int64_t kBf16Tile = 2048; // 32×32 × 2 B
+  static constexpr int64_t kBfp8Tile = 1088; // 32×32 × 1 B + 64 B row exponents
+  static constexpr int64_t kBfp4Tile =
+      576; // 32×32 × 0.5 B + 64 B row exponents
+  static constexpr int64_t kFp32Tile = 4096; // 32×32 × 4 B
+
+  int64_t kWeightTile =
+      (weightDataType == ttcore::DataType::BFP_BFloat4) ? kBfp4Tile : kBfp8Tile;
+
+  assert(p.kTiles % numIn0Cores == 0 &&
+         "kTiles must be divisible by numIn0Cores before the per-core divide");
+  int64_t kPerCore = p.kTiles / numIn0Cores;
+  // perCoreNCompute: tiles computed per DRAM-bank/compute core (= weight shard
+  // width per bank). Used for CB sizing — this is what the compute kernel
+  // actually accumulates per core before scattering to output storage cores.
+  int64_t perCoreNCompute = p.shardWTiles;
+
+  // Use numIn0Cores for the output tensor buffer estimate to keep the budget
+  // conservative and avoid inflating in0BlockW (which doubles in1CB per step).
+  // perCoreNStorage is only used for the output layout grid, not CB sizing.
+  int64_t outTensorBufPerCore =
+      p.perCoreM * ((N / kTileSize) / numIn0Cores) * kBf16Tile;
+  int64_t in0TensorBuf = p.perCoreM * kPerCore * kBf16Tile;
+  int64_t cbBudget = l1Available - in0TensorBuf - outTensorBufPerCore;
+
+  // Fixed CBs (independent of in0BlockW).
+  int64_t outCB = p.perCoreM * perCoreNCompute * kBf16Tile;
+  int64_t interm0CB = p.perCoreM * perCoreNCompute * kFp32Tile;
+  int64_t fixedCost = outCB + interm0CB;
+
+  if (fixedCost > cbBudget) {
+    return std::nullopt;
+  }
+
+  p.in0BlockW = kPerCore;
+  bool found = false;
+  while (p.in0BlockW >= 1) {
+    int64_t numBlocks = p.kTiles / p.in0BlockW;
+    bool doubleBuf = numBlocks > 1;
+
+    int64_t in0CB = p.in0BlockW * p.perCoreM * kBf16Tile * (doubleBuf ? 2 : 1);
+    int64_t in1CB = p.in0BlockW * perCoreNCompute * kWeightTile *
+                    (doubleBuf ? 3 : 1); // weight shard per DRAM bank
+
+    if (fixedCost + in0CB + in1CB <= cbBudget && kPerCore % p.in0BlockW == 0) {
+      found = true;
+      break;
+    }
+    if (p.in0BlockW == 1) {
+      break;
+    }
+    p.in0BlockW--;
+    while (p.in0BlockW > 1 && kPerCore % p.in0BlockW != 0) {
+      p.in0BlockW--;
+    }
+  }
+
+  if (!found) {
+    return std::nullopt;
+  }
+
+  return p;
+}
+
+TTNNLayoutAttr buildDRAMShardedWeightLayout(MLIRContext *ctx,
+                                            TTNNLayoutAttr origLayout,
+                                            const DRAMShardParams &p) {
+  auto startCoord = CoreCoordAttr::get(ctx, 0, 0);
+  auto endCoord = CoreCoordAttr::get(ctx, p.numBanks - 1, 0);
+  auto coreRange = CoreRangeAttr::get(ctx, startCoord, endCoord);
+  auto crs = CoreRangeSetAttr::get(ctx, {coreRange});
+  auto tileType =
+      ttcore::TileType::get(ctx, {kTileSize, kTileSize}, p.weightDataType);
+  auto dramSpace = BufferTypeAttr::get(ctx, BufferType::DRAM);
+  auto memrefType = MemRefType::get({p.kTiles, p.shardWTiles}, tileType,
+                                    MemRefLayoutAttrInterface{}, dramSpace);
+  auto memLayout =
+      TensorMemoryLayoutAttr::get(ctx, TensorMemoryLayout::WidthSharded);
+  return TTNNLayoutAttr::get(ctx, origLayout.getLinear(),
+                             llvm::ArrayRef<int64_t>{1, p.numBanks}, memrefType,
+                             memLayout, /*tensorMesh=*/nullptr,
+                             /*ignorePhysicalLayout=*/false, crs);
+}
+
+// Build an L1 width-sharded layout for `tensorShape` over `numCores`, using
+// canonical core placement that wraps across the worker grid. A single-row
+// placement (0,0)-(numCores-1,0) would be invalid once numCores exceeds the
+// grid width (e.g. 16 cores on an 8x8 grid) — validateTensorSpec rejects it —
+// so canonical placement fills the grid row-by-row (16 cores -> 8x2), mirroring
+// the DS output-layout construction.
+TTNNLayoutAttr buildL1ShardedLayout(MLIRContext *ctx, TTNNLayoutAttr origLayout,
+                                    llvm::ArrayRef<int64_t> tensorShape,
+                                    int64_t numCores,
+                                    ttcore::DeviceAttr deviceAttr) {
+  return TTNNLayoutAttr::Builder(origLayout, tensorShape)
+      .setBufferType(BufferType::L1)
+      .setMemoryLayout(
+          TensorMemoryLayoutAttr::get(ctx, TensorMemoryLayout::WidthSharded))
+      .setGridShape({1, numCores})
+      .buildWithCanonicalCorePlacement(deviceAttr);
+}
+
+MatmulMultiCoreReuseMultiCastDRAMShardedProgramConfigAttr
+buildDRAMShardedProgramConfig(MLIRContext *ctx, const DRAMShardParams &p,
+                              UnaryWithParamAttr fusedAct) {
+  return MatmulMultiCoreReuseMultiCastDRAMShardedProgramConfigAttr::get(
+      ctx, p.in0BlockW, p.perCoreM, p.perCoreN, fusedAct);
+}
+
+DeviceComputeKernelConfigAttr
+buildComputeConfig(MLIRContext *ctx, ttcore::DataType weightDataType) {
+  MathFidelity fidelity = (weightDataType == ttcore::DataType::BFP_BFloat4)
+                              ? MathFidelity::LoFi
+                              : MathFidelity::HiFi2;
+  return DeviceComputeKernelConfigAttr::get(
+      ctx,
+      /*mathFidelity=*/fidelity,
+      /*mathApproxMode=*/mlir::BoolAttr{},
+      /*fp32DestAccEn=*/mlir::BoolAttr::get(ctx, true),
+      /*packerL1Acc=*/mlir::BoolAttr::get(ctx, true),
+      /*dstFullSyncEn=*/mlir::BoolAttr{});
 }
 
 } // namespace mlir::tt::ttnn
